+++ src/deeponto/onto/ontology.py
@@ -45,6 +45,7 @@ from org.semanticweb.owlapi.apibinding import OWLManager  # type: ignore
 from org.semanticweb.owlapi.model import IRI, OWLObject, OWLClassExpression, OWLObjectPropertyExpression, OWLDataPropertyExpression, OWLIndividual, OWLAxiom, AddAxiom, RemoveAxiom, AxiomType  # type: ignore
 from org.semanticweb.HermiT import ReasonerFactory as HermitReasonerFactory  # type: ignore
 from org.semanticweb.elk.owlapi import ElkReasonerFactory  # type: ignore
+from org.apache.log4j import Logger, Level  # type: ignore
 from org.semanticweb.owlapi.util import OWLObjectDuplicator, OWLEntityRemover  # type: ignore
 from org.semanticweb.owlapi.search import EntitySearcher  # type: ignore
 
@@ -621,6 +622,7 @@ class OntologyReasoner:
             self.owl_reasoner_factory = HermitReasonerFactory()
         elif reasoner_type == "elk":
             self.owl_reasoner_factory = ElkReasonerFactory()
+            Logger.getLogger("org.semanticweb.elk").setLevel(Level.ERROR)
 
         self.owl_reasoner = self.owl_reasoner_factory.createReasoner(self.onto.owl_onto)
         self.owl_data_factory = self.onto.owl_data_factory