+++ docs/changelog.md
@@ -14,8 +14,7 @@ Security in case of vulnerabilities. -->
 - [X] **Add** new reasoner type `"struct"` ([Structural Reasoner](https://owlcs.github.io/owlapi/apidocs_4/org/semanticweb/owlapi/reasoner/structural/StructuralReasoner.html)) at `deeponto.onto.OntologyReasoner`.
 - [X] **Add** `load_reasoner()` method at `deeponto.onto.OntologyReasoner` for convenience of changing the reasoner type.
 - [x] **Add** `rdflib` into the dependencies for building graph-related features.
-- [ ] **Add** `deeponto.onto.OntologyTaxonomy` for building the subsumption graph from an ontology.
-- [X] **Add** `deeponto.onto.WordnetTaxonomy` for building the noun hypernym graph from wordnet (not an ontology but a commonly used KG).
+- [X] **Add** `deeponto.onto.taxonomy` for building the taxonomy over ontologies and potentially other structured data.
 
 ### Changed
 - [X] **Change** the default lowcasing argument of `deeponto.utils.process_annotation_literal()` to `False`.